simple topmenu added with the navigator button
